requirements.txt
aiogram==3.*
yt-dlp
python-dotenv
aiohttp
yt-dlp[tiktok]
